LayoutTests/inspector/debugger/properties-special-expected.txt

Tests how debugger presents special properties of closures, bound functions and object wrappers.

Debugger was enabled.
Script source was shown.
Set timer for test function.
Watch expressions updated.
Nodes are expanded.
error: Function scope data is not available
Subnodes are expanded.
'Object(true)' => 'Boolean'
    '__proto__' => 'Boolean'
'(function(a,b) { return a + b; })' => 'function (a, b) { return a + b; }'
    'arguments' => 'null'
    'caller' => 'null'
    'length' => '2'
    'name' => '""'
    'prototype' => 'Object'
        'constructor' => 'function (a, b) { return a + b; }'
        '__proto__' => 'Object'
    '__proto__' => 'function () {'
    <function scope>
'(function(a,b) { return a + b; }).bind({}, 2)' => 'function () {'
    'arguments' => null
        'get' => 'function () {'
        'set' => 'function () {'
    'caller' => null
        'get' => 'function () {'
        'set' => 'function () {'
    'length' => '1'
    'name' => '""'
    '__proto__' => 'function () {'
    <function scope>
Debugger was disabled.
